6.1 6.2 amma sduallarim var muellim((

+++ src/Core/OnionPronia.Application/DTOS/GetProductInCategoryDto.cs
@@ -4,6 +4,6 @@ namespace OnionPronia.Application.DTOS
 {
     public record GetProductInCategoryDto
          (
-             int Id, string Name, decimal Price
+             long Id, string Name, decimal Price
          );
 }
